+++ test/api/NewAddResourceAPISpec.js
@@ -15,7 +15,7 @@ describe('mongoose collection name API', function () {
   before(function (done) {
     app = express();
 
-    fng = new (formsAngular)(app, {urlPrefix: '/api/'});
+    fng = new (formsAngular)(mongoose, app, {urlPrefix: '/api/'});
 
     mongoose.connect('localhost', 'forms-ng_test');
     mongoose.connection.on('error', function () {