+++ test/api/reportAPISpec.js
@@ -16,7 +16,7 @@ describe('Report API', function () {
   before(function (done) {
     app = express();
 
-    fng = new (formsAngular)(app, {urlPrefix: '/api/'});
+    fng = new (formsAngular)(mongoose, app, {urlPrefix: '/api/'});
 
     mongoose.connect('localhost', 'forms-ng_test');
 //    mongoose.set('debug',true);